file_upload: Fix IndexError
https://github.com/ThreeSixtyGiving/insights-ng/issues/36
